backend/requirements.txt
watchdog==4.0.2
fastapi==0.114.0
starlette==0.38.5
uvicorn==0.30.6
python-dotenv==1.0.1